Updates
Reverted due to Project Errors
Recleaned Project (temp)

+++ HAPI_Start/Visualisation.h
@@ -25,7 +25,7 @@ public:
 	void BlitAlpha(BYTE* screen, Rectangle screenRect, Texture* texture, int posX, int posY, int CamX, int CamY); //Blits the Texture pixel by pixel (WITH ALPHA)
 	void loadTexture(std::string tName, bool hasAlpha = true); //pushes the texture into the SpriteMap
 	
-	int spriteMapSize() { return SpriteMap.size(); }
+	size_t spriteMapSize() { return SpriteMap.size(); }
 	//MOVE THIS TO PRIVATE
 	Texture* getSprite(int i); //returns the Texture of the specified index in the SpriteMap
 	std::unordered_map<int, Texture*>& getSpriteMap(); //Returns the SpriteMap
